[Navbar] Migrate the navbar to react-boostrap.
